Embed VPN management view
